Changes to sketches master and slave and to the speaker driver
Changed completely the method to set the volume of the M5Atom Echo speaker. For this reason I removed from the AtomEchoSPKR driver function setVolume because that method did not work. Now using a method in which the beep tone.maxval can be changed between 0 and 10. This works OK. Added a new Monitor output text file